Use fetch_user() when get_user() fails
get_user() is extremely unreliable for getting users because it relies 
on the bot's cache users for finding users. This cache appears to very 
limited in scope thus causing most users to be missed when trying the 
fetch them. fetch_user() will directly ping Discord servers to get more 
information about the user thus causing users to not be missed.