Added ping_devs to utils.py
Note: This function works different than the ping_devs method in 
bdaybot_commands.py. This function is now only used for pinging the devs 
as opposed to logging and returning strings

This function does two things
• Sends developers detailed error information via DM message (who 
receives the message can be configured before running the bot)
• Bot sends message in the channel where the error took place and 
mentions developers in the message

Logging must now be done in the function that calls ping_devs. Logging 
no longer occurs in ping_devs to make all the instances of logging more 
clear, even if it means additional lines of code.