Created housekeeping.py and CosmicHouseKeepingCog
This cog is for keeping the database in an as usable condition as 
possible. Additionally, when certain aspects of the bot no longer 
function this cog will let the guild owner know and sets on how to fix 
it.

Here are situations this cog handles
• Adds new guilds the bot has joined to the database
• Deletes non-existent announcements channels from database
• Deletes announcements channels the bot can no longer send messages in 
from the database
• Automatically re-adds Happy Birthday/Upcoming role when removed
• Notifies guild owner when roles are missing that prevent the bot from 
operating correctly
• Deletes announcements channel from database if deleted from guild