Fixed bug where the nickname wouldn't properly change
This bug was due of improper updating within SQLAlchemy. Whenever the 
today_name_cycler attribute is updated (via the next function), 
SQLAlchemy did not detect this change.  Even when trying to assign the 
cycler to a new variable and reassigning that new variable back to the 
today_name_cycle attribute it still didn't update due to an improperly 
implemented __eq__ method in itertools.cycle.  To get around this a 
comparator is manually defined in PickleType which always returns 
False whenever today_name_cycle column is updated. This means whenever a 
new value is assigned to the today_name_cycle attribute SQLAlchemy will 
update that value in the SQL database (when .commit() is called)