Added support for wishing people with hyphens in their names
This was really an edge case. A certain person in the database had a 
hyphen in their name which caused them not to be found in the database 
due to how the strings were processed. Before this change, .title() was 
used to process strings before using them to search the database, since 
all the names in database start with an uppercase letter.  However, 
.title() treats a hyphen like a space when converting letters to 
uppercase. For example, 'hello-world' is converted to 'Hello-World' 
which is undesirable behavior. With this commit, .capitalize() is now 
used to process the strings since it will only capitalize the first 
letter. As an example of it's behavior, 'hello-world' is converted to 
'Hello-world'.