Fixed another edge case with bdays on Jan 1st
Since all data in the DataFrame still uses the previous year when 
computing timedelta() all results are negative (since all these dates 
are in the past). This commit adds a check to make sure that the year 
for the data in the DataFrames is still the current year. If the current 
year and the years in the DataFrame are different then they will be 
updated to the current year.