Changed the way the delay for change_roles works
Instead of adding the delay in the loop like before, which usually did 
not work, the delay is now in a method that is invoked before the loop 
is started. This is how this function used to work and it had no issues 
with this implementation. Issues only started to appear with new 
implementation where the delay was in the loop itself.