Figured out why .get() was not working in test_wish and fixed the issue
When SQLAlchemy searches for a row in table when using .get() SQLAlchemy 
checks for both type equality (isinstance) and normal equality (==). 
This was an issue since wishee_id was a numpy.int64 type and not a 
normal Python integer. This caused SQLAlchemy to fail to find the Wish 
we are searching for. This commit converts wishee_id to a Python integer 
before using it in .get().

An interesting note is that when using .get() to retrieve data from a 
table that only has one primary key, SQLAlchemy does not check for type 
equality and only checks for normal equality. However, when searching in 
a table with multiple primary keys SQLAlchemy checks for both type 
equality and normal equality. This is why this issue only occured 
when accessing data from the wishes table and not when accessing data 
from the student_data table.