Reverted SQLAlchemy version back to 1.4.0b2
As is denoted in sqlalchemy#5920 there is currently no support for async SQLite3.  Our bdaybot HEAVILY relies on sqlite3 when testing so we desperately need this support.  In version 1.4.0b2 there is no check for whether or not the DBAPI is an async compatible.  Due to this, we can use the regular sync sqlite3 module when making transaction to an SQLite database.  This is actually a bug, however, since as of this commit, there is no support for SQLite databases within the async framework we must use this version of SQLAlchemy to enable all our connections to any SQLite3 database.  As soon as async SQLite is implemented we will immediately switch over the latest version.