Completely changed how the bdaybot distributes sessions
Previously every Cog had its own single session object that all the methods, tasks, and commands within that Cog would use. Although this structure, generally speaking, has worked, it has proven to be problematic in some situations.  For example when mutliple methods in a cog where reading and writing with the session sometimes their transactions would be melded in ways that would cause errors or write data that may not have been intented to be written. Now that every time a method is called a new session is created this will never happen.