Drastically improved the relability of the unit tests
Previously, sometimes the unit tests would fail even when they were supposed to pass.  This was due to an occasional error when the message used to trigger the command was mistaken for command response. However, now this commit implements the wait_for function which will wait until a message of a specific type has been sent until moving onto to the next line of code. This means there will never be an instance in which the unit tests fail unless there is a legitimate issue with the code.